+++ lib/createWebsite/makeRealWorldTable.js
@@ -2,8 +2,8 @@ import { writeFileSync } from 'node:fs';
 import { join } from 'node:path';
 
 import prettyMilliseconds from 'pretty-ms';
-import getRealWorldCSS from 'real-world-css-libraries';
 
+import { loadAllRealWorldTests } from '../loaders/loadAllRealWorldTests.js';
 import { minifiers, minify } from '../minify.js';
 
 const __dirname = import.meta.dirname;
@@ -11,34 +11,7 @@ const ERROR = 'ERROR';
 
 const minifyAllRealWorldTests = async function () {
   const startAll = Date.now();
-  let libraries = [];
-  try {
-    libraries = getRealWorldCSS();
-  } catch (error) {
-    console.log('Failed to read the CSS files.');
-    console.log(error);
-  }
-
-  libraries = libraries.filter((library) => {
-    // At this size limit these tests take around a minute,
-    // without this they take hours.
-    const fileIsReasonablySized = library.size < 500_000;
-
-    const knownBads = [
-      // Sass warns about deprecated @moz-document
-      'GitHub-Windows',
-      // Sass complains about incorrect hsla() syntax
-      'Halfmoon',
-      // Sass warns about deprecated darken()
-      'Jupyter Themes'
-    ];
-    const fileIsKnownBad = knownBads.includes(library.name);
-
-    return (
-      fileIsReasonablySized &&
-      !fileIsKnownBad
-    );
-  });
+  const libraries = loadAllRealWorldTests();
 
   for (const minifier of minifiers) {
     for (const library of libraries) {
